+++ src/data/IProduct.java
@@ -1,8 +1,10 @@
 package data;
 
+import java.math.BigDecimal;
+
 public interface IProduct {
     public String getSKU(); // Primary Key
     public String getName();
-    public double getPrice();
+    public BigDecimal getPrice();
     public IDistributor getDistributor(); // Foreign Key
 }
